refactor: new domain-driven folder structure
