+++ src/server/domain/config/repo/configRepository.ts
@@ -1,6 +1,6 @@
 import { env } from "~/env.mjs";
+import { type WidgetConfig } from "~/server/domain/config/widgetConfig";
 import { FileReader } from "~/server/driver/Reader/FileReader";
-import { type WidgetConfig } from "~/server/entities/widgetConfig";
 import AppError from "~/utils/error";
 import { ConfigLocalFileRepository } from "./configLocalFileRepository";
 import { ConfigRepositoryMock } from "./configRepositoryMock";
