+++ src/server/domain/config/widgetConfig.ts
@@ -3,14 +3,14 @@ import { isObject, isString } from "~/utils/guards/base";
 import { isFittingDataPaylod, isWidgetType } from "~/utils/guards/widgets";
 import { widgetTypeSchema } from "~/utils/schema";
 import type { WidgetType } from "~/utils/types/widget";
-import { calendarWidgetConfigSchema } from "../widgets/calendar/schema";
-import type { CalendarWidgetConfig } from "../widgets/calendar/types";
-import { linkWidgetConfigSchema } from "../widgets/links/schema";
-import type { LinkWidgetConfig } from "../widgets/links/types";
-import { searchWidgetConfigSchema } from "../widgets/search/schema";
-import type { SearchWidgetConfig } from "../widgets/search/types";
-import { timeWidgetConfigSchema } from "../widgets/time/schema";
-import type { TimeWidgetConfig } from "../widgets/time/types";
+import { calendarWidgetConfigSchema } from "../../widgets/calendar/schema";
+import type { CalendarWidgetConfig } from "../../widgets/calendar/types";
+import { linkWidgetConfigSchema } from "../../widgets/links/schema";
+import type { LinkWidgetConfig } from "../../widgets/links/types";
+import { searchWidgetConfigSchema } from "../../widgets/search/schema";
+import type { SearchWidgetConfig } from "../../widgets/search/types";
+import { timeWidgetConfigSchema } from "../../widgets/time/schema";
+import type { TimeWidgetConfig } from "../../widgets/time/types";
 
 export class WidgetConfig {
   id: string;
