+++ src/server/domain/layout/adjustedWidgetLayout.ts
