+++ src/server/domain/layout/services/adjustLayoutValuesService.ts
@@ -3,7 +3,7 @@ import { isScreenSize } from "~/utils/guards/other";
 import { isHidingInfo } from "~/utils/guards/widgets";
 import type { ScreenSize } from "~/utils/types/types";
 import type { HidingInfo, Positioning, WidgetType } from "~/utils/types/widget";
-import { type AdjustedWidgetLayout } from "../entities/adjustedWidgetLayout";
+import { type AdjustedWidgetLayout } from "../adjustedWidgetLayout";
 import {
   getMinHeightForWidget,
   getMinWidthForWidget,
