+++ src/server/domain/layout/services/computeSizeForWidgetService.test.ts
