+++ src/server/domain/layout/services/computeSizeForWidgetService.ts
