+++ src/server/domain/layout/services/transformWidgetLayoutService.ts
@@ -1,7 +1,7 @@
 import AppError from "~/utils/error";
 import { generateUniqueID } from "~/utils/helper";
-import { AdjustedWidgetLayout } from "../entities/adjustedWidgetLayout";
-import type { UserWidgetLayout } from "../entities/userWidgetLayout";
+import { AdjustedWidgetLayout } from "../adjustedWidgetLayout";
+import type { UserWidgetLayout } from "../userWidgetLayout";
 import addMissingLayouts from "./addMissingLayoutsService";
 import adjustLayoutValues from "./adjustLayoutValuesService";
