+++ src/server/widgets/calendar/service/dateManipulationService.test.ts
@@ -3,7 +3,7 @@
  */
 
 import type { CalendarComponent, FullCalendar } from "ical";
-import type { CalendarEntry } from "../widgets/calendar/types";
+import type { CalendarEntry } from "../types";
 import {
   getCalendarComponentMock,
   getRecurringCalendarComponentMock,
